Have check-it-out also appear under position B in mobile.
