+++ app/assets/stylesheets/better_homepage/application.css.sass
@@ -6,6 +6,20 @@
   // width: 100%
   z-index: 100
 
+#check-it-out
+  &-wide
+    @media (min-width: $media-mobile)
+      display: block
+    display: none
+    ul
+      @extend .c-list--horiz
+  &-block
+    @media (max-width: $media-mobile)
+      display: block
+    display: none
+    ul
+      @extend .c-list--vert
+
 section#content
 
   // Unsure why this couldn't be nested.
