+++ app/views/better_homepage/_contents.html.erb
@@ -22,6 +22,9 @@
         <a class="c-ad__attribution" href="http://www.scpr.org/support/underwriting/">Become a KPCC Sponsor</a>
       </div>
     </aside>
+    <div id='check-it-out-block'>
+      <%= precached "layouts/better_homepage/check_it_out" %>
+    </div>
   <% elsif index == 2 %>
     <div id='whats-next-initial' data-components='whats-next' class='boundary' data-scroll-up='hide' data-scroll-down='show' data-offset='-400'></div>
   <% elsif index == 7 %>
